# Reduce memory usage in image processing pipeline