--- a/imgutil/scale_test.go
+++ b/imgutil/scale_test.go
@@ -12,24 +12,41 @@ import (
 
 var genGoldenFiles = flag.Bool("gen_golden_files", false, "whether to generate the TestXxx golden files.")
 
-func TestKernelScaler(t *testing.T) {
+func TestScaler(t *testing.T) {
 	tests := []struct {
-		name  string
-		w     int
-		h     int
-		image string
+		name   string
+		w      int
+		h      int
+		image  string
+		scaler imgutil.Scaler
 	}{
 		{
-			name:  "downscale",
-			w:     100,
-			h:     100,
-			image: "wikipe-tan-100x123",
+			name:   "CatmullRom-downscale",
+			w:      100,
+			h:      100,
+			image:  "wikipe-tan-100x123",
+			scaler: imgutil.CatmullRom,
 		},
 		{
-			name:  "upscale",
-			w:     130,
-			h:     150,
-			image: "wikipe-tan-100x123",
+			name:   "CatmullRom-upscale",
+			w:      130,
+			h:      150,
+			image:  "wikipe-tan-100x123",
+			scaler: imgutil.CatmullRom,
+		},
+		{
+			name:   "CacheCatmullRom-downscale",
+			w:      100,
+			h:      100,
+			image:  "wikipe-tan-100x123",
+			scaler: imgutil.NewCacheScaler(imgutil.CatmullRom),
+		},
+		{
+			name:   "CacheCatmullRom-upscale",
+			w:      130,
+			h:      150,
+			image:  "wikipe-tan-100x123",
+			scaler: imgutil.NewCacheScaler(imgutil.CatmullRom),
 		},
 	}
 	for _, tt := range tests {
@@ -38,7 +55,7 @@ func TestKernelScaler(t *testing.T) {
 			goldenFname := fmt.Sprintf("testdata/%s-%s.png", tt.image, tt.name)
 
 			got := image.NewGray(image.Rect(0, 0, tt.w, tt.h))
-			imgutil.CatmullRom.Scale(got, src)
+			tt.scaler.Scale(got, src)
 
 			if *genGoldenFiles {
 				if err := writeImg(goldenFname, got); err != nil {
@@ -55,13 +72,50 @@ func TestKernelScaler(t *testing.T) {
 	}
 }
 
-func BenchmarkKernelScaler(b *testing.B) {
-	src := mustBeGray(mustReadImg("testdata/wikipe-tan-Gray.png"))
-	dstRect := imgutil.FitRect(src.Bounds(), 150, 150)
-	scaler := imgutil.CatmullRom.NewScaler(dstRect.Dx(), dstRect.Dy(), src.Rect.Dx(), src.Rect.Dy())
-	b.ResetTimer()
-	for i := 0; i < b.N; i++ {
-		dst := image.NewGray(dstRect)
-		scaler.Scale(dst, src)
+// BenchmarkScaler benchmarks scalers against one or more images. The occurences of each image are
+// evenly distributed among all benchmark iterations.
+func BenchmarkScaler(b *testing.B) {
+	benchmarks := []struct {
+		name   string
+		scaler imgutil.Scaler
+		images []string
+	}{
+		{
+			name:   "CatmullRom_singleImage",
+			scaler: imgutil.CatmullRom.NewScaler(122, 150, 195, 239),
+			images: []string{"wikipe-tan-195x239.png"},
+		},
+		{
+			name:   "CatmullRom_threeImages",
+			scaler: imgutil.CatmullRom.NewScaler(122, 150, 195, 239),
+			images: []string{"wikipe-tan-195x239.png", "wikipe-tan-100x123.png", "wikipe-tan-82x100.png"},
+		},
+		{
+			name:   "CacheCatmullRom_singleImage",
+			scaler: imgutil.NewCacheScaler(imgutil.CatmullRom),
+			images: []string{"wikipe-tan-195x239.png"},
+		},
+		{
+			name:   "CacheCatmullRom_threeImages",
+			scaler: imgutil.NewCacheScaler(imgutil.CatmullRom),
+			images: []string{"wikipe-tan-195x239.png", "wikipe-tan-100x123.png", "wikipe-tan-82x100.png"},
+		},
+	}
+	for _, bb := range benchmarks {
+		b.Run(bb.name, func(b *testing.B) {
+			var images []*image.Gray
+			for _, i := range bb.images {
+				images = append(images, mustBeGray(mustReadImg("testdata/"+i)))
+			}
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				b.StopTimer()
+				img := images[i%len(images)]
+				dstRect := imgutil.FitRect(img.Bounds(), 150, 150)
+				b.StartTimer()
+				dst := image.NewGray(dstRect)
+				bb.scaler.Scale(dst, img)
+			}
+		})
 	}
 }