… imgutil/imgutil_test.go
… imgutil/testdata/wikipe-tan-100x123-CacheCatmullRom-downscale.png
… imgutil/testdata/wikipe-tan-100x123-CacheCatmullRom-upscale.png
… imgutil/testdata/wikipe-tan-100x123-CatmullRom-downscale.png
… imgutil/testdata/wikipe-tan-100x123-CatmullRom-upscale.png
… imgutil/testdata/wikipe-tan-195x239.png
… imgutil/testdata/wikipe-tan-82x100.png