@naisuuuu 2021-04-19
Some progress was made.
Step by step (left to right oldest to newest commit):
```
name \ time/op          f5df3ac.txt  d288da7.txt  2e4a668.txt  a5816d6.txt
Converter/grayscale-12   8.30s ± 1%   8.06s ± 1%   8.07s ± 2%   8.08s ± 1%
Converter/ycbcr-12       10.6s ± 1%   10.1s ± 1%   10.3s ± 1%   10.2s ± 1%
Converter/testdata-12   7.05ms ± 1%  7.00ms ± 2%  6.83ms ± 4%  6.83ms ± 2%

name \ alloc/op         f5df3ac.txt  d288da7.txt  2e4a668.txt  a5816d6.txt
Converter/grayscale-12  10.4GB ± 0%   8.2GB ± 3%   6.9GB ± 4%   6.0GB ± 6%
Converter/ycbcr-12      28.7GB ± 0%  24.9GB ± 2%  23.5GB ± 2%  23.6GB ± 1%
Converter/testdata-12   7.47MB ± 0%  6.34MB ± 1%  5.60MB ± 2%  5.29MB ± 1%

name \ allocs/op        f5df3ac.txt  d288da7.txt  2e4a668.txt  a5816d6.txt
Converter/grayscale-12   24.1k ± 2%   22.8k ± 2%   22.7k ± 2%   22.3k ± 2%
Converter/ycbcr-12       6.61M ± 0%   6.61M ± 0%   6.61M ± 0%   6.61M ± 0%
Converter/testdata-12    1.89k ± 0%   1.82k ± 0%   1.83k ± 0%   1.80k ± 0%
```
Cumulative (old @ f5df3ac, new @ a5816d6):
```
name                    old time/op    new time/op    delta
Converter/grayscale-12     8.30s ± 1%     8.08s ± 1%   -2.65%  (p=0.008 n=5+5)
Converter/ycbcr-12         10.6s ± 1%     10.2s ± 1%   -3.29%  (p=0.008 n=5+5)
Converter/testdata-12     7.05ms ± 1%    6.83ms ± 2%   -3.09%  (p=0.032 n=5+5)

name                    old alloc/op   new alloc/op   delta
Converter/grayscale-12    10.4GB ± 0%     6.0GB ± 6%  -42.16%  (p=0.008 n=5+5)
Converter/ycbcr-12        28.7GB ± 0%    23.6GB ± 1%  -17.67%  (p=0.008 n=5+5)
Converter/testdata-12     7.47MB ± 0%    5.29MB ± 1%  -29.12%  (p=0.008 n=5+5)

name                    old allocs/op  new allocs/op  delta
Converter/grayscale-12     24.1k ± 2%     22.3k ± 2%   -7.32%  (p=0.008 n=5+5)
Converter/ycbcr-12         6.61M ± 0%     6.61M ± 0%   -0.02%  (p=0.008 n=5+5)
Converter/testdata-12      1.89k ± 0%     1.80k ± 0%   -4.93%  (p=0.008 n=5+5)
```
The above needs some explanation.

`Converter/testdata` benchmarks downscaling and converting the contents of `imgutil/testdata` to fit a 150x150px bounding box. This data set is available in this repository.
`Converter/grayscale` and `Converter/ycbcr` reflect a more realistic usage scenario - they benchmark downscaling and converting two entries of my manga collection (entire volumes) to fit a 1920x1920px bounding box. I cannot commit this data set into this repository, as I am not legally allowed to distribute it.

The difference in performance between those benchmarks is due to a few factors. First off, the data set size:
data set | number of files | size | archive format | avg resolution | target resolution (bounding box)
-----------|----------------------|-------|--------------------|---------------------|----------------------
testdata|16                       |327k|dir                    |195x239           |150x150
ycbcr    |200                     |290m|zip                   |2146x3056       |1920x1920
grayscale|189                  |303m|zip                   |1725x2475       |1920x1920

Although the ycbcr and grayscale data sets appear similar, profiling shows one major difference between them. The grayscale set is mostly decoded directly into `image.Gray`, whereas the ycbcr set is mostly decoded into `image.YCbCr`. There is some impact of this on our conversion pipeline code of course, but the biggest allocations by far come from the stdlib's `jpeg.Decode` (62.5% in grayscale dataset, 82.7% in ycbcr) The following are graph visualizations of memory allocation size in both cases: [ycbcr-memprofile.pdf](https://github.com/naisuuuu/mangaconv/files/6332748/ycbcr-memprofile.pdf), [grayscale-memprofile.pdf](https://github.com/naisuuuu/mangaconv/files/6332754/grayscale-memprofile.pdf).

With the overwhelming majority of memory allocated coming directly from stdlib implementation, I'm left with a few choices for further optimization, none of which seem particularly appealing:
- First, I could simply fork the stdlib and narrow it's implementation to optimize for this specific usage. This is far from ideal, as it requires a sizeable copy-paste and most likely a considerable amount of work.
- Second, suppose there is a solution to make the decoder allocate less memory that is generic enough to fit the current `jpeg.Decoder` api. This could potentially be a good change to the stdlib itself, however I am not at all knowledgeable enough about jpeg decoders to be confident I could come up with something better than the current implementation. Even if I could this would be, again, a considerable amount of work.
- Third, cgo. Using a libvips wrapper or similar would no doubt increase performance and improve memory usage. It would however also mean runtime dependencies. That's a bit of a bummer and something I embarked upon this project specifically meaning to avoid.
- Fourth, unused memory is wasted memory. Most systems should have enough to handle running the current implementation, right?

I don't immediately see a best options among the above.