@naisuuuu 2021-04-19
The only easy optimization left here that I can think of is handling an edge case in which an image is decoded as `*image.Paletted`. So far this has happened once and is visible as the massive amount of allocations in the `Converter/ycbcr` benchmark from the comment above. This is enough of an edge case with little impact on performance that I'm going to leave it on the table.

Regarding to the previous comment, I think I may proceed with the third option (introducing a runtime dependency) in the future. For now, the tool is usable and fairly quick despite the excessive volume of allocations.